requirements.txt
aiohttp==3.5.4
appnope==0.1.0
astroid==2.2.5
async-timeout==3.0.1
attrs==19.1.0
backcall==0.1.0
chardet==3.0.4
decorator==4.4.0
discord.py==1.2.3
idna==2.8
idna-ssl==1.1.0
ipython==7.8.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
lazy-object-proxy==1.4.2
mccabe==0.6.1
multidict==4.5.2
numpy==1.17.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
pylint==2.3.1
python-dotenv==0.10.3
six==1.12.0
traitlets==4.3.2
typed-ast==1.4.0
typing-extensions==3.7.4
wcwidth==0.1.7
websockets==6.0
wrapt==1.11.2
yarl==1.3.0
